Source-MC->string exchange through SC success
About Source-MC,
With this code built in MC, it can accept a string sent to it through SC
and echo it back to the sender.
The string can be of variable length but its end must be marked by a null ascii character.
For eg - to send "Hello", send "Hello\0". \0 is the null charater.

To work with variable length character arrays I have implemented the string struct to allow for most
basic functionality that cpp string library offers. This can be improved in further commits.